refactor: table 관련 공통 컴포넌트 utils로 이동
